valuehistogram: new simple, value-based histogram
This change adds a new valuehistogram panel to support value-based
histograms.  Unlike the previous histogram panel which plots the data
using fixed time-interval buckets, this type of histogram buckets the
data according to the values themselves.  This type of graph provides
a high-level perspective of a field's value across the entire time
interval.