Fixes and Visual Improvements, support for Missing/Exists, self-documenting panels feature and more (#12)
* US16124: ressurect Missing values feature from Terms Panel

On Terms Panels (Terms mode, no Term_Stats or Unique Counts), the original option to compute and show the count of docs where the field is null or missing now works again

* US16124: tighten up stats panel to accurately leverage the stat format option

Previously, if the stat panel was to show the Count of docs, it would still format it as currency/bytes/etc, which doesn't make any sense and could be misleading especially for bytes

* US16124: fix regression rendering chart slices if labels are numeric values

* US16124: fix other issue with numeric values being aggregated in terms mode for multi-select filtering operation

* US16124: add panel help/description option to help report viewers understand the info they see

Terms, Histo, Ranges and Stats were implemented (not all). Should be a smarter global way to do this, figured to research more later but get the value out now

* US16124: Extend Hits panel to better control display of  featured stats (number format and text align)

* US16124: implement progressive disclosure of panel config options

Reduce cognitive load for report viewers who rarely need to interact with this command bar that increases clutter. TODO: mobile/no-hover devices get the CTAs always, make that better if we ever need to better support

* US16124: upgrade dark/light mode themes

Modernized font, improve readability and contrast in light mode

* US16124: missed font size for dark mode

* US16124: first round of grunt build errors fixes

including upgrading jshint to support modern javascript

* US16124: round 2 jshint fixes, lot of orig old stuff to retrofit

* US16124: add tmp folder to git ignore setup

not sure why grunt build created this now, unless it has to do with the upgraded jshint package. set to ignore for now

* US16124: round 3 build fixes and package upgrades, v7.1.0

Replace grunt-ngmin with grunt-ng-annotate, couple other dependencies. clean build now. increment version to 7.1 and deploy to test server to verify